refactor: Autorização nas rotas

+++ Tryitter/Tryitter/Controllers/PostController.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 using Tryitter.Repository;
 
 namespace Tryitter.Controllers;
@@ -15,6 +16,7 @@ public class PostController : Controller
         _repository = repository;
     }
     [HttpGet]
+    [AllowAnonymous]
     public IActionResult Get()
     {
         var posts = _repository.GetPosts();
@@ -22,14 +24,25 @@ public class PostController : Controller
     }
     
     [HttpGet("{id}")]
+    [Authorize]
     public IActionResult GetById(int id)
     {
+        var resultId = _repository.GetOwnerPost(id);
+        if (resultId == null)
+        {
+            return NotFound($"Post {id} not Found");
+        }
+        else if (User.Identity.Name != Convert.ToString(resultId))
+        {
+            return Unauthorized("Usuário não logado");
+        }
         var posts = _repository.GetPost(id);
         if (posts == null) return NotFound("Post not exists");
         return Ok(posts);
     }
 
     [HttpPost]
+    [Authorize]
     public IActionResult Create(string content, int userId)
     {
         var createdPost = _repository.AddPost(content, userId);
@@ -38,16 +51,36 @@ public class PostController : Controller
     }
 
     [HttpDelete("{id}")]
+    [Authorize]
     public IActionResult Delete(int id)
     {
+        var resultId = _repository.GetOwnerPost(id);
+        if (resultId == null)
+        {
+            return NotFound($"Post {id} not Found");
+        }
+        else if (User.Identity.Name != Convert.ToString(resultId))
+        {
+            return Unauthorized("Usuário não logado");
+        }
         var isDeleted = _repository.DeletePost(id);
         if (isDeleted) return Ok();
         return BadRequest();
     }
 
     [HttpPut("{id}")]
+    [Authorize]
     public IActionResult Update(int id, string? content)
     {
+        var resultId = _repository.GetOwnerPost(id);
+        if (resultId == null)
+        {
+            return NotFound($"Post {id} not Found");
+        }
+        else if (User.Identity.Name != Convert.ToString(resultId))
+        {
+            return Unauthorized("Usuário não logado");
+        }
         var isUpdated = _repository.UpdatePost(id, content!);
         if (isUpdated) return Ok();
         return BadRequest();
